add pipe tests
