+++ pipe_test.go
@@ -1,5 +1,74 @@
 package pipe_test
 
+import (
+	"errors"
+	"testing"
+
+	"pipelined.dev/pipe"
+	"pipelined.dev/pipe/mock"
+)
+
+func TestLineBindingFail(t *testing.T) {
+	var (
+		errorBinding = errors.New("binding error")
+		bufferSize   = 512
+	)
+	testBinding := func(l pipe.Routing) func(*testing.T) {
+		return func(t *testing.T) {
+			t.Helper()
+			_, err := pipe.New(bufferSize, l)
+			if err == nil {
+				return
+			}
+			if !errors.Is(err, errorBinding) {
+				t.Fatal("error")
+			}
+		}
+	}
+	t.Run("source", testBinding(
+		pipe.Routing{
+			Source: (&mock.Source{
+				ErrorOnMake: errorBinding,
+			}).Source(),
+			Processors: pipe.Processors(
+				(&mock.Processor{}).Processor(),
+			),
+			Sink: (&mock.Sink{}).Sink(),
+		},
+	))
+	t.Run("processor", testBinding(
+		pipe.Routing{
+			Source: (&mock.Source{}).Source(),
+			Processors: pipe.Processors(
+				(&mock.Processor{
+					ErrorOnMake: errorBinding,
+				}).Processor(),
+			),
+			Sink: (&mock.Sink{}).Sink(),
+		},
+	))
+	t.Run("sink", testBinding(
+		pipe.Routing{
+			Source: (&mock.Source{}).Source(),
+			Processors: pipe.Processors(
+				(&mock.Processor{}).Processor(),
+			),
+			Sink: (&mock.Sink{
+				ErrorOnMake: errorBinding,
+			}).Sink(),
+		},
+	))
+	t.Run("ok", testBinding(
+		pipe.Routing{
+			Source: (&mock.Source{}).Source(),
+			Processors: pipe.Processors(
+				(&mock.Processor{}).Processor(),
+			),
+			Sink: (&mock.Sink{}).Sink(),
+		},
+	))
+}
+
 // func TestInsertProcessor(t *testing.T) {
 // 	bufferSize := 2
 // 	testInsert := func(pos int) func(*testing.T) {
